Reduce to single history graph source
